backend/pyproject.toml
[tool.poetry]
name = "oraculo-backend"
version = "0.1.0"
description = "Backend API for the Oraculo AI Workspace Platform"
authors = ["Manus <seu-email@exemplo.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.111.0"}
celery = "^5.4.0"
redis = "^5.0.4"
sqlalchemy = "^2.0.30"
supabase = "^2.4.3"
google-generativeai = "^0.5.4"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pydantic-settings = "^2.3.1"
psycopg2-binary = "^2.9.9"
pymupdf = "^1.24.5"
python-docx = "^1.1.2"
openpyxl = "^3.1.3"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.1"
ruff = "^0.4.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"